Pull out some helpers from mcp/views/[viewId]/index.ts
